Updated api

+++ src/controllers/playlist.controller.js
@@ -5,10 +5,19 @@ const Boom = require('@hapi/boom');
 
 playlistController = {};
 
-playlistController.listAll = async () => {
+playlistController.listTop = async () => {
     try {
         const playLists = await PlaylistModel.find();
-        return(playLists);
+        return (playLists);
+    } catch (error) {
+        return BoomError(Boom.badRequest(error));
+    }
+}
+
+playlistController.listAll = async ({ userEmail }) => {
+    try {
+        const playLists = await PlaylistModel.find({owner: userEmail});
+        return (playLists);
     } catch (error) {
         return BoomError(Boom.badRequest(error));
     }
@@ -20,24 +29,23 @@ playlistController.getPlaylist = async (id) => {
         if (!playlist) {
             return BoomError(Boom.notFound('Playlist not found'));
         }
-        return (playlist);    
+        return (playlist);
     } catch (error) {
         return BoomError(Boom.badRequest(error));
     }
 }
 
-playlistController.createPlaylist = async (playlist) => {
+playlistController.createPlaylist = async ({ body, userEmail }) => {
+    const playlist = body;
+    playlist.owner = userEmail;
     try {
-
         const validation = PlaylistSchema.validate(playlist);
-
         if (validation.error) {
             error = validation.error.details[0].message;
             return BoomError(Boom.badRequest(error));
         }
-
         const newPlaylist = PlaylistModel.create(playlist);
-        return (newPlaylist);   
+        return (newPlaylist);
     } catch (error) {
         return BoomError(Boom.badRequest(error));
     }
@@ -46,24 +54,27 @@ playlistController.createPlaylist = async (playlist) => {
 playlistController.updatePlaylist = async (id, playlist) => {
     try {
         const validation = PlaylistSchema.validate(playlist);
-
         if (validation.error) {
             error = validation.error.details[0].message;
             return BoomError(Boom.badRequest(error));
         }
-
-        const updatedPlaylist = await PlaylistModel.findByIdAndUpdate(id, playlist, {new: true});
+        const updatedPlaylist = await PlaylistModel.findByIdAndUpdate(id, playlist, { new: true });
         if (!updatedPlaylist) {
             return BoomError(Boom.notFound('Playlist not found'));
         }
         return (updatedPlaylist);
-    } catch(error) {
+    } catch (error) {
         return BoomError(Boom.badRequest(error));
     }
 }
 
-playlistController.removePlaylist = async (id) => {
+playlistController.removePlaylist = async ({params, userEmail}) => {
+    const id = params.id;
     try {
+        const playlist = await PlaylistModel.findOne({_id: id});
+        if (playlist.owner !== userEmail) {
+            return BoomError(Boom.unauthorized('This is not your playlist'));
+        }
         const removedPlaylist = await PlaylistModel.findByIdAndDelete(id);
         if (!removedPlaylist) {
             return BoomError(Boom.notFound('Playlist not found'));
@@ -77,17 +88,18 @@ playlistController.removePlaylist = async (id) => {
 playlistController.addExercise = async (id, exercise) => {
     try {
         const playlist = await PlaylistModel.findOne(
-            {'_id': id, 
-            exercises: {'$in': [exercise._id]}}
+            {
+                '_id': id,
+                exercises: { '$in': [exercise._id] }
+            }
         );
-
         if (playlist) {
             return BoomError(Boom.badRequest('Exercise already exists in playlist'));
         } else {
             const updatedPlaylist = await PlaylistModel.findByIdAndUpdate(
-                {'_id': id},
-                {$push: {'exercises': exercise._id}},
-                {new: true}
+                { '_id': id },
+                { $push: { 'exercises': exercise._id } },
+                { new: true }
             );
             return (updatedPlaylist);
         }
@@ -99,15 +111,16 @@ playlistController.addExercise = async (id, exercise) => {
 playlistController.removeExercise = async (id, exercise) => {
     try {
         const playlist = await PlaylistModel.findOne(
-            {'_id': id, 
-            exercises: {'$in': [exercise._id]}}
+            {
+                '_id': id,
+                exercises: { '$in': [exercise._id] }
+            }
         );
-
         if (playlist) {
             const updatedPlaylist = await PlaylistModel.findByIdAndUpdate(
-                {'_id': id},
-                {$pullAll: {'exercises': [exercise._id]}},
-                {new: true}
+                { '_id': id },
+                { $pullAll: { 'exercises': [exercise._id] } },
+                { new: true }
             );
             return (updatedPlaylist);
         } else {
